předělávka struktury u /YYYYMMDD/metrics dle Tomových komentářů v Pull Requestu no.1
