+++ src/main/java/cz/centrum/haffner/SimpleTrainingDavid/REST/SimpleTrainingDavidApplication.java
@@ -1,9 +1,14 @@
 package cz.centrum.haffner.SimpleTrainingDavid.REST;
 
 import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
 
-@SpringBootApplication
+@Configuration
+@EnableAutoConfiguration
+@ComponentScan ("cz.centrum.haffner.SimpleTrainingDavid")
 public class SimpleTrainingDavidApplication {
 
 	public static void main(String[] args) {
