+++ src/test/java/cz/centrum/haffner/SimpleTrainingDavid/SimpleTrainingDavidApplicationTests.java
@@ -1,9 +1,10 @@
 package cz.centrum.haffner.SimpleTrainingDavid;
 
 import org.junit.jupiter.api.Test;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.test.context.SpringBootTest;
 
-
+@SpringBootApplication
 @SpringBootTest
 class SimpleTrainingDavidApplicationTests {
